moved slide to components
