+++ src/components/New/FrontPage.tsx
@@ -1,35 +1,30 @@
 import styled, { keyframes } from 'styled-components';
+import background1D from '../../assets/images/homepage/page1/background1D.png';
+import background2D from '../../assets/images/homepage/page2/background2D.png';
+import background3D from '../../assets/images/homepage/page3/background3D.png';
+import background1M from '../../assets/images/homepage/page1/background1M.png';
+import background2M from '../../assets/images/homepage/page2/background2M.png';
+import background3M from '../../assets/images/homepage/page3/background3M.png';
 
+const backgroundD = [background1D, background2D, background3D];
+const backgroundM = [background1M, background2M, background3M];
 interface SlidesProps {
   translate: any;
 }
-export const SlidesDesktop = styled.div<SlidesProps>`
+export const Slides = styled.div<SlidesProps>`
   display: flex;
   flex-direction: row;
   width: 300vw;
   transition: transform 1s;
   transform: ${(props) => `translateX(${props.translate}%)`};
-  @media screen and (max-width: 600px) {
-    display: none;
-  }
 `;
-export const SlidesMobile = styled.div<SlidesProps>`
-  display: flex;
-  flex-direction: row;
-  width: 300vw;
-  transition: transform 1s;
-  transform: ${(props) => `translateX(${props.translate}%)`};
-  @media screen and (min-width: 601px) {
-    display: none;
-  }
-`;
+
 interface BackgroundWrapperProps {
-  background: string;
   translate: any;
   index: number;
 }
 export const BackgroundWrapper = styled.div<BackgroundWrapperProps>`
-  background-image: url(${(props) => props.background});
+  background-image: url(${(props) => backgroundD[props.index]});
   background-position: ${(props) =>
     props.index == 0 ? 'center right' : props.index == 1 ? 'center right 30%' : 'center'};
   background-size: cover;
@@ -41,5 +36,6 @@ export const BackgroundWrapper = styled.div<BackgroundWrapperProps>`
   overflow: hidden;
   @media screen and (max-width: 601px) {
     background-position: center center;
+    background-image: url(${(props) => backgroundM[props.index]});
   }
 `;
